test - organize unit tests into separate modules
Use one test module for each code module being tested.
This lets us separate our tests to keep them clean and organized.
