+++ test/__init__.py
@@ -1,0 +1,6 @@
+"""
+Network Notary Server unit tests.
+Tests are separated into one module of tests for each module of code being tested.
+
+Run unit_tests.py to run all tests.
+"""
